Finish WelcomeView
WelcomeView now creates appropriate sets of classes divided by level and section. Classes are stored as an array of Class. these are then stored in a dictionary for each level of type [Int:[Class]] where the Int is the section number. these are then stored in a dictionary [String:[Int:[Class]]] which is passed to the admin Controller.