Change how the main view displays the sections
Instead of the section being in each column the sections are now the rows. upon selecting a section(clicking the row on the table) it is now displayed in the table on the side. Also added a check on the welcome view to make sure they have the correct number of sections for the level 6 classes.