Filter time and room lists for selection
Filters the times list based on when the selected teacher is available. Filters the room based on what time is selected.
Fixed some faulty checking in WelcomeViewController for having set all level 6 classes.
Fixed error in PreferenceOverrideController causing it to crash from not yet having the teachers array when the table was being generated